:hammer: Move PDARule to new file
